chore(go-server): rename internal to pkg
this enables external users to reuse our logic with their own custom main and other functions. Future work will make auth and other features more pluggable
