+++ packages/server/duckdb-server-go/pkg/query/options.go
