+++ packages/server/duckdb-server-go/pkg/query/query.go
