+++ packages/server/duckdb-server-go/pkg/query/validator.go
