+++ packages/server/duckdb-server-go/pkg/query/validator_test.go
