+++ packages/server/duckdb-server-go/pkg/server/server.go
@@ -12,7 +12,7 @@ import (
 	"github.com/coder/websocket"
 	"github.com/coder/websocket/wsjson"
 
-	"github.com/uwdata/mosaic/packages/server/duckdb-server-go/internal/query"
+	"github.com/uwdata/mosaic/packages/server/duckdb-server-go/pkg/query"
 )
 
 type Command string
